Typecheck typescript tests again
