+++ spec/typescript/middleware/csrf.spec.ts
@@ -1,10 +1,9 @@
 import forge from 'mappersmith'
 import CSRF from 'mappersmith/middleware/csrf'
 
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-const client = forge({
-  middleware: [ CSRF('csrfToken', 'x-csrf-token') ],
+forge({
+  middleware: [CSRF('csrfToken', 'x-csrf-token')],
   clientId: 'github',
   host: 'https://status.github.com',
-  resources: {}
+  resources: {},
 })
