+++ spec/typescript/middleware/global-error-handler.spec.ts
@@ -6,10 +6,9 @@ setErrorHandler((response) => {
   return response.status() === 500
 })
 
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-const client = forge({
-  middleware: [ GlobalErrorHandler ],
+forge({
+  middleware: [GlobalErrorHandler],
   clientId: 'github',
   host: 'https://status.github.com',
-  resources: {}
+  resources: {},
 })
