+++ spec/typescript/middleware/log.spec.ts
@@ -1,10 +1,9 @@
 import forge from 'mappersmith'
 import Log from 'mappersmith/middleware/log'
 
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-const client = forge({
-  middleware: [ Log ],
+forge({
+  middleware: [Log],
   clientId: 'github',
   host: 'https://status.github.com',
-  resources: {}
+  resources: {},
 })
